add namespace to issuer, use kingdonb everywhere
Since we only have access to one namespace, and CRDs are preinstalled
for us, things are actually much easier for us.

Use the kingdonb namespace everywhere, and change any cluster-wide
resources to not be cluster-wide. We will get access to a privileged
helm-operator later for other namespaces, that can be installed
by a cluster-admin or somehow otherwise out-of-band.

Also, upgrade to latest SealedSecrets chart release.